ml_service/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
albumentations>=1.3.0
timm>=0.9.0
torchcam>=0.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.4.0
prometheus-fastapi-instrumentator>=6.1.0
aiofiles>=23.2.0
